Changed Method Of HACluster Password And Corosync Auth Key
Changed the method of defining the HACluster user
password. The previous method was continually
regenerating the password and syncing up all of the
hosts.

Changed the method in which to capture the
Corosync Auth Key. The previous method used cat
to capture the auth key and always triggered the
task(s) as a change. Now using slurp has resolved
this issue.

Signed-off-by: Larry Smith Jr <mrlesmithjr@gmail.com>